feat: sfml display

+++ src/shared/IGraphical.hpp
@@ -9,13 +9,14 @@
 #define IGRAPHICAL_HPP_
 
 #include <utility>
+#include <vector>
 #include "../static_libs/math/Color.hpp"
 
 class IGraphical {
     public:
         ~IGraphical() = default;
         virtual void setSize(const std::pair<int, int> &size) = 0;
-        virtual void drawPixel(int x, int y, const math::Color &color) = 0;
+        virtual void drawPixel(size_t x, size_t y, const math::Color &color) = 0;
 };
 
 #endif /* !IGRAPHICAL_HPP_ */
